TODO: loading logic should recurse up directory parents and replace them with asterisks
